WTO1016/WTO_moni/project_new/requirements.txt
# ========================================
# WTO会议系统依赖包
# ========================================

# Web框架核心
Flask==2.3.3                    # Web应用框架
Werkzeug==2.3.7                 # WSGI工具库
Flask-CORS==4.0.0               # 跨域资源共享支持

# 认证和会话管理
Flask-Login==0.6.3              # 用户登录管理
Flask-JWT-Extended==4.5.3       # JWT认证
bcrypt==4.0.1                   # 密码加密

# 实时通信
Flask-SocketIO==5.3.6           # WebSocket支持
python-socketio==5.9.0          # Socket.IO客户端
eventlet==0.33.3                # 异步服务器

# 数据库
pymongo==4.5.0                  # MongoDB驱动

# 文件处理
PyPDF2==3.0.1                   # PDF文本提取
python-docx==1.1.2              # Word文档处理
reportlab==4.0.7                # PDF生成

# 文本处理和AI
jieba==0.42.1                   # 中文分词
scikit-learn==1.4.1             # 机器学习和文本相似度
numpy==1.26.3                   # 数值计算
scipy==1.12.0                   # 科学计算

# HTTP请求和环境变量
requests==2.31.0                # HTTP客户端
python-dotenv==1.0.0            # 环境变量管理

# 日期时间处理
python-dateutil==2.8.2          # 日期时间工具

# 系统兼容性
importlib-metadata==6.8.0       # Python模块元数据（兼容性）

# ========================================
# 安装说明
# ========================================
# 安装所有依赖：
# pip install -r requirements.txt
#
# 如果安装速度慢，可以使用国内镜像：
# pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple/
#
# 验证安装：
# python -c "import flask, pymongo, PyPDF2; print('依赖安装成功')"
# ========================================